backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
google-cloud-vision==3.7.4
snowflake-connector-python==3.11.0
httpx==0.27.2
websockets==13.1
python-multipart==0.0.9
lxml==5.3.0
Pillow==10.4.0
deepdiff==7.0.1

# Watcher dependencies
watchdog==4.0.1
pygetwindow==0.0.9
